Fix negative values on stock status Report

+++ modules/report/src/main/java/org/openlmis/report/model/report/VaccineStockStatusReport.java
@@ -40,7 +40,7 @@ public class VaccineStockStatusReport implements ResultRow {
 
     private Long isaValue;
 
-    private Integer mos;
+    private Float mos;
 
     //private JSONPObject products;
